Organize CSS and improved element transition from dark to light
I added a few more CSS classes and organized the properties (order goes:
background/color, display properties, width/height, margin/padding, font
properties, and the rest of the properties in alphabetical order with
transform being last).

For some reason, applying the dark class to each individual icon on the
bottom bar doesn't work in Mobile (tried on iPhone Safari and Firefox),
but it works on desktop Firefox and Chromium.