Fix image border on mobile and add minimal Readme
On an iPhone, the filter CSS property would not take the border-radius
into effect, causing a border to appear around the image during the
animation. Now, the clip-path property is used. If that property is not
supported on a web browser, it will default to the same behavior as
before. I also made the image circular by default so, in case clip-path
isn't supported, the image will still be round.